+++ lib/models/iexpense.dart
@@ -1,9 +1,9 @@
-class Todo {
+class IExpense {
   Resource resource;
 
-  Todo({this.resource});
+  IExpense({this.resource});
 
-  Todo.fromJson(Map<String, dynamic> json) {
+  IExpense.fromJson(Map<String, dynamic> json) {
     resource = json['resource'] != null
         ? new Resource.fromJson(json['resource'])
         : null;
@@ -19,35 +19,38 @@ class Todo {
 }
 
 class Resource {
-  List<Data> data;
+  Reff ref;
+  int ts;
+  Data data;
 
-  Resource({this.data});
+  Resource({this.ref, this.ts, this.data});
 
   Resource.fromJson(Map<String, dynamic> json) {
-    if (json['data'] != null) {
-      data = new List<Data>();
-      json['data'].forEach((v) {
-        data.add(new Data.fromJson(v));
-      });
-    }
+    ref = json['ref'] != null ? new Reff.fromJson(json['ref']) : null;
+    ts = json['ts'];
+    data = json['data'] != null ? new Data.fromJson(json['data']) : null;
   }
 
   Map<String, dynamic> toJson() {
     final Map<String, dynamic> data = new Map<String, dynamic>();
+    if (this.ref != null) {
+      data['ref'] = this.ref.toJson();
+    }
+    data['ts'] = this.ts;
     if (this.data != null) {
-      data['data'] = this.data.map((v) => v.toJson()).toList();
+      data['data'] = this.data.toJson();
     }
     return data;
   }
 }
 
-class Data {
-  Reff ref;
+class Reff {
+  Refff ref;
 
-  Data({this.ref});
+  Reff({this.ref});
 
-  Data.fromJson(Map<String, dynamic> json) {
-    ref = json['@ref'] != null ? new Reff.fromJson(json['@ref']) : null;
+  Reff.fromJson(Map<String, dynamic> json) {
+    ref = json['@ref'] != null ? new Refff.fromJson(json['@ref']) : null;
   }
 
   Map<String, dynamic> toJson() {
@@ -59,16 +62,16 @@ class Data {
   }
 }
 
-class Reff {
+class Refff {
   String id;
-  Data collection;
+  Reffff collection;
 
-  Reff({this.id, this.collection});
+  Refff({this.id, this.collection});
 
-  Reff.fromJson(Map<String, dynamic> json) {
+  Refff.fromJson(Map<String, dynamic> json) {
     id = json['id'];
     collection = json['collection'] != null
-        ? new Data.fromJson(json['collection'])
+        ? new Reffff.fromJson(json['collection'])
         : null;
   }
 
@@ -82,18 +85,37 @@ class Reff {
   }
 }
 
-class Refff {
+class Reffff {
   String id;
 
-  Refff({this.id});
+  Reffff({this.id});
 
-  Refff.fromJson(Map<String, dynamic> json) {
+  Reffff.fromJson(Map<String, dynamic> json) {
     id = json['id'];
   }
 
   Map<String, dynamic> toJson() {
     final Map<String, dynamic> data = new Map<String, dynamic>();
     data['id'] = this.id;
+    return data;
+  }
+}
+
+class Data {
+  String name;
+  String amount;
+
+  Data({this.name, this.amount});
+
+  Data.fromJson(Map<String, dynamic> json) {
+    name = json['name'];
+    amount = json['amount'];
+  }
+
+  Map<String, dynamic> toJson() {
+    final Map<String, dynamic> data = new Map<String, dynamic>();
+    data['name'] = this.name;
+    data['amount'] = this.amount;
     return data;
   }
 }
