+++ lib/widgets/expenseCard.dart
@@ -1,19 +1,15 @@
-import 'package:fexpense/models/itodo.dart';
+import 'package:fexpense/models/iexpense.dart';
 import 'package:flutter/material.dart';
 
 typedef DeleteCallback = Function(String id);
 
 class ExpenseCard extends StatelessWidget {
-  final ITodo todo;
-  // final String content, id;
+  final IExpense expense;
   final DeleteCallback delete;
-  // final TextEditingController updateText = TextEditingController();
 
   ExpenseCard(
       {Key key,
-      @required this.todo,
-      // @required this.content,
-      // @required this.id,
+      @required this.expense,
       @required this.delete})
       : super(key: key);
 
@@ -27,17 +23,17 @@ class ExpenseCard extends StatelessWidget {
           color: Colors.red,
         ),
         title: Text(
-          todo.resource.data.name,
+          expense.resource.data.name,
           style: TextStyle(
               fontSize: 15, fontWeight: FontWeight.bold, color: Colors.grey),
         ),
         subtitle: Text(
-          "Amount:" + todo.resource.data.amount,
+          "Amount:" + expense.resource.data.amount,
           style: TextStyle(fontSize: 15, color: Colors.grey),
         ),
         trailing: IconButton(
             icon: Icon(Icons.delete, color: Colors.redAccent, size: 20.0),
-            onPressed: () => delete(todo.resource.ref.ref.id)),
+            onPressed: () => delete(expense.resource.ref.ref.id)),
       ),
     );
   }
